mb: Add support for modifying a simplechrome builder's GN args.
And utilize it to add the arg "ozone_platform_headless=true" to the fyi
tester and run ozone_gl_unittests.

Context: There are a subset of tests that prefer to run with the
"headless" ozone platform (see http://crbug.com/859246#c1).

The chrome-chromeos ebuild (and by association, the cros chrome-sdk)
whitelists the ozone platforms that are supported:
https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/45c432cbf1fac8cf655dd5c28c543097fdaefd43/chromeos-base/chromeos-chrome/chromeos-chrome-9999.ebuild#356

For the amd64-generic board, it turns off auto_platform detection, and
whitelists only GBM. Consequently, using those GN args it generates,
we can't run in headless mode on our bots.

So we need to whitelist 'headless' mode. This is done by adding the
"ozone_platform_headless=true" GN arg. We can do this by either:
- Adding it to the ebuild.
Or
- Making MB add it during "mb.py gen" time.

This chooses the latter option because the ebuild's configuration is
(FWIU) optimized for building a production grade version of the Chrome
binary on ChromeOS. The 'headless' mode of ozone is (also FWIU) more of
debugging/test mode and isn't suitable for prod builds of Chrome. So
barring any "debug" mode of an ebuild that I'm not aware of, this goes
with the mb option.

Bug: 860066
Change-Id: I07234c67c2e493375525da227b8754de2a20fdb7
Reviewed-on: https://chromium-review.googlesource.com/1142740
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#576536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74ad5377d3cbe059f589b388deea6a077d8e2c2d