Add intermediate bubble UI for local card migration flow (CL#3).
This is the 3rd CL adding metrics and unit-tests.

Previous CL: 1128569, 1134193.

For project details, please refer to design doc go/ib-paradise.

This project is to offer users an option to upload chrome local credit
cards to Google Pay server. This bubble UI is the entry point of the entire
flow. It should be anchored to a credit card icon in the location bar and
be shown from it.

Bug: 852904
Change-Id: I286a32d9673eb6341a9e577055068bba93d93b42
Reviewed-on: https://chromium-review.googlesource.com/1135932
Commit-Queue: Siyu An <siyua@chromium.org>
Reviewed-by: Robert Kaplow (OOO until 0730) <rkaplow@chromium.org>
Reviewed-by: Mathieu Perreault <mathp@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#576674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87cc15bf6de998097879793c0249d3b74ffd8e4f