[TTS] And Unity integration about:flag feature.
Adds a new Feature for integration with Unified Consent.
When this is enabled, and unified consent is on, the TTS
opt-in promo will not be needed.  We'll implement a follow-on
CL that does the conditional disabling off of this flag.

BUG=825428

Change-Id: I4becefb6bf3d1862f941bccd85c2bbc863cb99d4
Reviewed-on: https://chromium-review.googlesource.com/1148689
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#577708}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e9b5b7fba690f6c98f3c612a7cd0d06da469531d