[VK] Add private extension API to set virtual keyboard hit-test rects.
In the new virtual keyboard UI, the floating virtual keyboard has a
an irregular shape. This is implemented by drawing the shape over a
transparent background, so it looks like the window has the correct
shape. However, clicking the transparent areas does not propagate to
the window behind the keyboard, breaking the illusion.

We add a new private extension API for setting a custom hit test area
for the virtual keyboard from the IME extension. By using this API, the
IME can specify the exact shape of the virtual keyboard so that events
are propagated correctly based on the shape.

Only the new UI will be calling this API, so nothing changes for the
old UI (hit test area == window bounds).

Bug: 866380
Change-Id: I38d9cc7f263d09cb3343be6d298667bdba7dd9ea
Reviewed-on: https://chromium-review.googlesource.com/1147680
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#578143}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 489344842fc67d3c006ee95c70b35a8fff285234