[Password Manager] Migrate contents of native backends into LoginDB
PasswordStoreX now takes 3 backends: a native backend, an unencrypted
LoginDB and an encrypted LoginDB. If the unencrypted LoginDB has data
(because the native backend was previously unavailable), its contents
will be moved into the native backend. If that succeeds, the contents
of the native backend will be moved into the encrypted LoginDB. From
then on, PasswordStoreX will only use the encrypted LoginDB. Progress
is tracked on a user preference.

This migration may take a non-trivial amount of time, therefore the
steps are designed to be safe to repeat if Chrome fails: the source
of the data is not cleared. Also, the encrypted LoginDB is a different
location than the unencrypted one, since distinguishing the two just
from their contents is unreliable.

In a future version, we will clear Chrome's data from the native
backend and move the encrypted LoginDB onto the standard location.

The migration is disabled behind a feature. Enabling the feature will
perform the migration once. If the feature is disabled again, the
store will serve again from the (possibly out-of-date) native backend.
Re-enabling will not repeat a migration for any new data produced while
disabled.

Bug: 571003
Change-Id: I2206b043638ae1484b30ac884c5cf2d0acaea2b8
Reviewed-on: https://chromium-review.googlesource.com/1120165
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Christos Froussios <cfroussios@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#579424}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e060c4c274d71369a74e0ecb6e8eef677bd3584