Add one-time IPH for bottom toolbar
This patch adds one-time IPH for the bottom toolbar for the first time
that it is enabled. The help bubble appears under the bottom toolbar
and states that the icons have moved to the bottom of the screen.

Bug: 862611
Change-Id: I597401a1030cc9f0c4ea3537bcf9a597b8d28275
Reviewed-on: https://chromium-review.googlesource.com/1138729
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Robert Kaplow (slow) <rkaplow@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#580255}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7d9ba3b289a917c1b61e3982a8ae41e061e1bf2