[Power ML] Change smart dim model interface to return ModelPrediction.
Model could error out so we need to record this response in the log.
Also, the logged inactivity_score and decision_threshold are quantized
version of the actual values used in the inference. Hence the
model now returns it decision as part of the prediction.

Bug: 862461
Change-Id: I14c889d97bebd93b88a86027ff94b2ce7fad5bfb
Reviewed-on: https://chromium-review.googlesource.com/1154243
Commit-Queue: Jia Meng <jiameng@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#580740}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b3f0a775610500d24e12f1d00a3cc1787de7bdb6