mb: Set symbol_level to 1 on Mac/Win sanitizer builders.
The intent of crrev.com/579533 was to use the same symbol_level on
sanitizer bots as we did before crrev.com/579330 landed by using the
default symbol level on all sanitizer bots. That was thought to be 1
on every platform (which would match the previous behaviour), but it
turns out that on Mac/Win, the default symbol_level on sanitizer bots
is 2, not 1, so the effect of crrev.com/579330 was to increase the
symbol level on those bots and crrev.com/579533 had no effect. This
change implements the intent of crrev.com/579533 by setting the symbol
level back to 1.

Bug: 869361
Change-Id: I33e572b215b389fb7d0201f5e6d11415530d86b3
Reviewed-on: https://chromium-review.googlesource.com/1164532
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#581138}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f08052f98dee02f9ec6912ce3050e41143ee599