SuperSize: Improve logic for when to omit symbols in html_report.py
It was using a hardcoded MAX_SYMBOLS. It now will keep adding symbols
until 95% of the size is attributed.

This increases the size of ndjson files for a monochrome .size file, but
should reduce the size of diffs.

Change-Id: I56146b20af43b0a47209943ea03bce98d562281f
Reviewed-on: https://chromium-review.googlesource.com/1176177
Commit-Queue: agrieve <agrieve@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#583496}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3c5b0b54680e5ff40efbfa388d3ed2f10eb633ac