[Autofill Views] Adding Layout Experiment
This CL adds the flag, feature, and parameters for an upcoming
experiment which will modify the layout of the Autofill dropdown on
Desktop. It also adds the capability to display the icon on the left,
rather than the default right, of each row of the dropdown, which is
the expected behavior under certain experiment groups.

Change-Id: I90ceb546887a76eb836e4aac19ae96ba1a9d107e
Bug: 870342
Reviewed-on: https://chromium-review.googlesource.com/1152217
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#584572}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3fa3df5cdc0cfa96f7c46a1d837e3e1e8e16b2da