Add a UMA to track the size of JPEG images.
This CL introduces the Blink.ImageDecoders.Jpeg.Area UMA which tracks
the area (total # of pixels) of JPEG images. The rationale is that the
area can be used as a proxy for how useful a JPEG decode accelerator
will be: the larger the area, the more likely an accelerator is useful.

Bug: 868400
Change-Id: I5252dea60107cd5160e01970b8f6c490c0931b9e
Reviewed-on: https://chromium-review.googlesource.com/1172914
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Leon Scroggins <scroggo@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#586491}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e43eef10c4b17dc1705682639a1b25c3fecb9d86