Add histogram for the reason of the rollback
Send a histogram value from UpgradeDetector when the device did
install a rollback image and shows a notification before reboot.

The metric records whether the rollback is happening because of a
channel switch or because of an admin-initiated enterprise
rollback.

Bug: 871759
Change-Id: I38fad1792ba87b6bdbd788b3cd0f736dfa2d363c
Reviewed-on: https://chromium-review.googlesource.com/1190703
Commit-Queue: Marton Hunyady <hunyadym@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#586724}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8f26669fb2113ff56e033fc6604d92dc7c1b803f