Add ServiceWorkerImportedScriptUpdateCheck flag
This CL tentatively adds ServiceWorkerImportedScriptUpdateCheck flag.
It is for the implementation of byte-for-byte update check for
service worker importScripts() resources. We are going to implement
the feature under this flag and flip it over once it is completed.

Bug: 648295
Change-Id: Iabb6140af84382ecd0d07ee28283490c303e8c4d
Reviewed-on: https://chromium-review.googlesource.com/1192571
Commit-Queue: Momoko Hattori <momohatt@google.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#587017}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fc3522926869bf9cbaca9111eaacebaf412ad038