Add histograms for time to update.
This change adds histograms for the total time from when the update is
first seen by the client to when the update is downloaded (but before
the device has rebooted). There are two histograms, one for the total
time in minutes when the client has a time restriction policy for
updates, and one for when the client has no policy in place.

These two histograms are only reported for enterprise devices.

CL where the histogram is used:
https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/1179230


BUG=chromium:852860
TEST=Run through scenarios and checked chrome://histograms

Change-Id: Ib4b8d0126087d5951b79d7f4d2022c2c776744a7
Reviewed-on: https://chromium-review.googlesource.com/1181649
Commit-Queue: May Lippert <maybelle@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#587752}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ab1e95f63731525e5ea3340cddf6bbdea278ed1d