[signin] Metrics for ListAccounts success rate
It is suspected (from external data) that calls to Gaia in general, and
ListAccounts in particular, have an abnormally high network failure rate.

This CL adds Chrome metrics to measure the failure rate of ListAccounts,
to see if the phenomenon can be observed through Chrome metrics:
"Gaia.AuthFetcher.ListAccounts.NetErrors"

Additional histograms are added to track the failure rate of ListAccounts
depending on system uptime and process uptime, because one of the
hypothesis is that network requests are generally more likely to fail on
system or process startup, and ListAccounts requests tend to be done at
startup.

The Signin.ListAccounts.Failure histogram now also records success, so
that it can be used to compute failure rates.

"Gaia.AuthFetcher.Multilogin.AuthErrors" is also added to track success
rate for OAuth2 multilogin, which needs to be monitored as it is ramped
up.

Bug: 876306
Change-Id: I3feb3a48ba43b79559fe1aacd78ce51eb0b86a38
Reviewed-on: https://chromium-review.googlesource.com/1195463
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#589136}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 86a630eb36591d1e3e383eeaa80317882dd24f00