Add a metric to track omnibox warmup
Some omnibox janks are related to singleton not yet initialized.

We believe there can be lot of code running under the hood on the first
key stroke. We observed font cache initialisation, localisation loading, etc...

The intend of this metrics is to be able to compare popupation for the
first key stroke and for the normal use of omnibox.

R=mpearson@chromium.org

Bug: 868419
Change-Id: I191fd6ee564aa584b633b66a2f90fc7752b64633
Reviewed-on: https://chromium-review.googlesource.com/1210167
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#589640}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a1fb72cf2b0ec35d1922fd47a453536cf53d0757