Add UMA metric for Shill sandboxing.
Chrome OS metric signifying whether a system is running Shill in a
sandbox or not. This is controlled through a Finch flag, but requires
1-2 reboots to take affect. This metric will help us ensure that Shill
is being run with sandboxing enabled/disabled at the frequency that we
expect.

Bug: 649417
Change-Id: Ia88980e813b6c633996f7964c9eb35deb041d248
Reviewed-on: https://chromium-review.googlesource.com/1210862
Commit-Queue: Micah Morton <mortonm@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#590393}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9eda51db13b743b778146730762f08eb1ad71ae6