Creates a new dialog for Crostini Upgrade.
If the user's cros-termina component is incompatible with the current version
of Chrome and must be upgraded, but the user is offline, we show this dialog
asking the user to go online and try again. This happens when they attempt to
launch any crostini app (e.g. the terminal). Testing the dialog showed a bug
in which the spinner for Terminal was not dismissed if the CrostiniRestart
failed before the standard spinner delay. This has been fixed in the current
CL.

Also added testing code to configure the component updater in browser tests so
we can simulate the circumstances leading to the dialog being shown.

Bug: 881616
Change-Id: I1007a3bd2c34c8403ab80335063072c4995d5bff
Reviewed-on: https://chromium-review.googlesource.com/1218264
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#590937}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc9185c627f678ff37431b6e9e112f72cabe0f4d