Orderfile: add logging on repeated symbols.
There are frequently a small number of symbols that are repeated with
different offsets. These are benign but cause some annoyances, for
example they can make it appear that a binary is not ordered correctly,
necessitating the threshold used in check_orderfile.py (q.v.).

This CL adds a small amount of logging for the repeated symbols that
will make it easier to see the cause if a larger number of symbols get
repeated.

Bug: 758566
Change-Id: I04038bebec38b1b38e5667a578e7b4e06ac6060d
Reviewed-on: https://chromium-review.googlesource.com/1201853
Reviewed-by: Benoit L <lizeb@chromium.org>
Commit-Queue: Matthew Cary <mattcary@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#591314}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 799a3b2f8c0ebd82b7940b446d1354057606b37c