[Supersize] Add a flag for uncompressed .pak files.
This CL adds a new flag for non-locale .pak files that aren't
compressed. This is exposed in the UI as a highlight option and will be
used by the size trybot.

This CL also changes the default unit selection to be KiB.

Bug: 702625
Change-Id: I7385b2b0c6897bb8c79cfd93eaa446789fd515a9
Reviewed-on: https://chromium-review.googlesource.com/1225593
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: agrieve <agrieve@chromium.org>
Commit-Queue: Eric Stevenson <estevenson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#591773}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e9cb2681888d293a4a5e12cdd51e9bf4991a620