Fix GeneratedCodeCache::ReadData to handle incomplete entries
In GeneratedCodeCache::ReadDataComplete also handle cases when
the entry was just created but no data is written into. We
always expect response_time as the header in the cache entry. If
the number of bytes read is smaller than the response_header then
it should be treated as a Error. Also, we can have entries with
only response time and no data associated with it.

Bug: chromium:884217
Change-Id: I5cc3f0923aa87ce208fe766ab6e9fb2633c24159
Reviewed-on: https://chromium-review.googlesource.com/1228123
Commit-Queue: Mythri Alle <mythria@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#592414}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4c2245a2453e29859b39f00a998c1e444833627