Fix UAF triggered by clearing MediaCapabilities DB
For incognito window's the in-memory DB is seeded (read only) by the DB
of the original profile. This is achieved via raw pointer to the seed
DB and the invariant that the incognito profile is always outlived by
its ancestor profile. Unfortunately, this overlooked that the ancestor
profile may destroy the DB and make a new one when asked to clear the
DB history. This caused the seed pointer to be invalid.

This CL changes the VideoDecodePerfHistory to never destroy the DB.
The DB is now a const unique_ptr, passed in during construction.

This CL also changes how clearing history works. Before, we used the
Destroy() API of the underlying ProtoDatabase to delete all entries.
This left the DB broken until re-initialized, which presents a risk
whenever that DB is seeding an incognito profile. Going forward,
clearing the DB is achieved by querying for all the keys and then
removing them. This removes the need for re-initialization and allows
the seeded in-memory DB to carry on safely unaware.

Test: Updated unit tests. Also manual e2e testing.
Bug: 881059
Change-Id: I014d76d9e04b16574fd7eea08775b1388740dd8d
Reviewed-on: https://chromium-review.googlesource.com/1220392
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#593659}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06b8109c05e05134a2c4235922b5c394e6c1a84d